packages/sdk/pyproject.toml
[tool.poetry]
name = "pureml"
version = "0.1.3.0"
description = ""
license = "Apache-2.0"
authors = ["vamsidhar muthireddy <vamsi.muthireddy@gmail.com>"]
readme = "README.md"
homepage = "https://pureml.com/"
repository = "https://github.com/engageml-github/Pure"
documentation = "https://docs.pureml.com"
keywords = ["pureml", "model-store", "machine-learning", "python", "model-registry", "collabortion"]
classifiers = [
    'Topic :: Software Development :: Build Tools',
    'Topic :: Software Development :: Libraries :: Python Modules',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
]



[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0"
pydantic = "^1.9.1"
numpy = "^1.23.1"
typer = {extras = ["all"], version = "^0.6.1"}
PyJWT = "^2.4.0"
requests = "^2.28.1"
python-dotenv = "^0.20.0"
Pillow = "^9.3.0"
pyarrow = "^8.0.0"
pandas = "^1.4.3"
matplotlib = "^3.6.2"
joblib = "^1.2.0"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
docker = "^6.0.1"

[tool.poetry.dev-dependencies]
ipykernel = "^6.19.2"
lightgbm = "^3.3.2"
xgboost = "^1.6.1"
catboost = "^1.0.6"
scikit-learn = "^1.1.1"
fastparquet = "^0.8.1"
pytest = "^7.1.3"
scipy = {version = "^1.9.1", python = "^3.8,<3.12"}
pandas-profiling = {version="^3.4.0",  python = "^3.8,<3.11"}
ipywidgets = "^8.0.2"
tensorflow = {version="^2.10.1",  python = "^3.8,<3.11"}
notebook = "^6.5.2"
scikit-image = "^0.19.3"
opencv-python = "^4.6.0"
nltk = "^3.7"
pytorch-tabnet = "^4.0"



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
pureml = "pureml.cli.main:app"


[tool.pytest.ini_options]
testpaths = [
    "tests",
    "tests/models",
    "tests/components"
]